requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
groq
bcrypt==4.0.1
python-jose[cryptography]
python-multipart